backend/requirements.txt
asgiref==3.6.0
astroid==2.15.4
atomicwrites==1.4.1
attrs==22.2.0
Brotli==1.0.9
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.1.0
colorama==0.4.6
coreapi==2.3.3
coreschema==0.0.4
cryptography==40.0.1
cssselect2==0.7.0
defusedxml==0.7.1
diff-match-patch==20200713
dill==0.3.6
Django==3.2.18
django-cors-headers==3.14.0
django-extra-fields==3.0.2
django-filter==23.1
django-import-export==3.1.0
django-templated-mail==1.1.1
django-weasyprint==2.2.0
djangorestframework==3.14.0
djoser==2.1.0
et-xmlfile==1.1.0
flake8==5.0.4
fonttools==4.38.0
gobject==0.1.0
gunicorn==20.1.0
html5lib==1.1
idna==3.4
importlib-metadata==1.7.0
iniconfig==2.0.0
itypes==1.2.0
Jinja2==3.1.2
lazy-object-proxy==1.9.0
MarkupPy==1.14
MarkupSafe==2.1.2
mccabe==0.7.0
mypy==1.2.0
mypy-extensions==1.0.0
oauthlib==3.2.2
odfpy==1.4.1
openpyxl==3.1.2
packaging==23.0
Pillow==9.5.0
platformdirs==3.3.0
pluggy==1.0.0
psycopg2-binary==2.9.5
py==1.11.0
pycodestyle==2.9.0
pycparser==2.21
pydyf==0.6.0
pyflakes==2.5.0
PyJWT==2.6.0
pyphen==0.14.0
pytest==6.2.5
pytest-pythonpath==0.7.4
python-dotenv==0.21.1
python3-openid==3.2.0
pytz==2023.3
PyYAML==6.0
requests==2.28.2
requests-oauthlib==1.3.1
six==1.16.0
social-auth-app-django==4.0.0
social-auth-core==4.4.1
sqlparse==0.4.3
tablib==3.4.0
tinycss2==1.2.1
toml==0.10.2
tomli==2.0.1
tomlkit==0.11.7
typing_extensions==4.5.0
uritemplate==4.1.1
urllib3==1.26.15
weasyprint==58.1
webencodings==0.5.1
wrapt==1.15.0
xlrd==2.0.1
xlwt==1.3.0
zipp==3.15.0
zopfli==0.2.2
